got rid of parsing helpers file
